feat: Update database schema for EnderecoCliente and Fornecedor
The commit updates the database schema for the `EnderecoCliente` and `Fornecedor` entities. The `create-tables.ts` file has been modified to change the data types of the `cep` and `numero` fields in the `EnderecoCliente` table from `INT` to `VARCHAR`. This change allows for more flexibility in storing the address information. Additionally, the `create-tables.ts` file has been updated to remove the `total_pedido` field from the `DetalhesDoPedido` table.

Co-authored-by: Thiago Trainotti <thisgot@users.noreply.github.com>