feat: add pedidos routes and controllers
The commit adds the `pedidosRoutes` and `pedidosControllers` to the application. These routes and controllers handle requests related to pedidos, including fetching all pedidos, fetching details of a specific pedido, fetching status of all pedidos, and creating a new pedido with details and status. This change enhances the functionality of the application by providing CRUD operations for pedidos and improves the overall user experience.

Co-authored-by: Thiago Trainotti <thisgot@users.noreply.github.com>